+++ StatelessCMS/MenuIcon.php
@@ -3,19 +3,21 @@
 namespace Stateless;
 
 /**
- * @brief A single item in a Menu, represented by a font awesome icon
+ * A single item in a Menu, represented by a font awesome icon
  */
 class MenuIcon extends MenuItem {
 
     /**
-     * @brief Output the menu item markup to the current output buffer
+     * Output the menu item markup to the current output buffer
      */
     public function show() {
+
         $this->label = 
             "<i class=\"" . $this->label . "\" " .
-            "aria-hidden=\"true\"></i>"
-        ;
+            "aria-hidden=\"true\"></i>";
 
         parent::show();
+
     }
-}
+
+};